requirements.txt
# Core RAG
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM
google-generativeai==0.3.2

# Document Processing
pypdf==4.0.1

# UI
gradio==4.16.0

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
tavily-python>=0.3.0
